requirements.txt
google-analytics-data
google-auth
